Trade_platform/requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.7
django-celery-beat==2.2.1
django-extensions==3.1.3
django-rest==0.8.7
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.3.2
flake8==3.9.2
flower==1.0.0
humanize==3.12.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
pipenv==2021.5.29
platformdirs==2.4.0
prometheus-client==0.8.0
prompt-toolkit==3.0.22
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
